game timer

+++ Kuma Engine/Application.h
@@ -6,6 +6,7 @@
 #include "Globals.h"
 #include "Memory.h"
 #include "Timer.h"
+#include "Time.h"
 #include "Module.h"
 #include "Parson/parson.h"
 #include "PerfTimer.h"
@@ -82,7 +83,7 @@ public:
 	JSON_Object* LoadJSONFile(const std::string &path);
 	void Log(const char* fmt,...);
 	void saveLog(const char* fmt,...);
-	Uint32 GetMsTimer();
+	float GetMsTimer();
 
 	const char* GetAppName() const;
 	void SetAppName(const char* name);
